fix(game): Phase1 - You can go to the lobby

+++ AAEmu.Game/Core/Managers/HousingManager.cs
@@ -65,7 +65,7 @@ public class HousingManager(
     /// <param name="values"></param>
     /// <param name="accountId"></param>
     /// <returns></returns>
-    public int GetByAccountId(Dictionary<uint, House> values, uint accountId)
+    public int GetByAccountId(Dictionary<uint, House> values, ulong accountId)
     {
         foreach (var (id, house) in _houses)
             if (house.AccountId == accountId)
@@ -723,7 +723,7 @@ public class HousingManager(
     /// <param name="hostileTaxRate"></param>
     /// <param name="oneWeekTaxCount"></param>
     /// <returns></returns>
-    public bool CalculateBuildingTaxInfo(uint accountId, HousingTemplate newHouseTemplate, bool buildingNewHouse, out int totalTaxToPay, out int heavyHouseCount, out int normalHouseCount, out int hostileTaxRate, out int oneWeekTaxCount)
+    public bool CalculateBuildingTaxInfo(ulong accountId, HousingTemplate newHouseTemplate, bool buildingNewHouse, out int totalTaxToPay, out int heavyHouseCount, out int normalHouseCount, out int hostileTaxRate, out int oneWeekTaxCount)
     {
         totalTaxToPay = 0;
         heavyHouseCount = 0;
