+++ AAEmu.Game/Core/Managers/World/IEnterWorldManager.cs
@@ -6,8 +6,8 @@ namespace AAEmu.Game.Core.Managers.World;
 
 public interface IEnterWorldManager
 {
-    void AddAccount(uint accountId, uint connectionId);
-    void Login(GameConnection connection, uint accountId, uint token);
+    void AddAccount(ulong accountId, uint connectionId);
+    void Login(GameConnection connection, ulong accountId, uint token);
     void Leave(GameConnection connection, LeaveWorldTargetType leaveWorldTargetType);
     void LeaveWorldTask(GameConnection connection, LeaveWorldTargetType leaveWorldTarget, Character activeChar);
 }
