+++ AAEmu.Game/Core/Managers/World/IStreamManager.cs
@@ -4,7 +4,7 @@ namespace AAEmu.Game.Core.Managers.World;
 
 public interface IStreamManager
 {
-    void AddToken(uint accountId, uint connectionId);
+    void AddToken(ulong accountId, uint connectionId);
     void RemoveToken(uint token);
-    void Login(StreamConnection connection, uint accountId, uint token);
+    void Login(StreamConnection connection, ulong accountId, uint token);
 }
