+++ AAEmu.Game/Core/Packets/C2S/CTJoinPacket.cs
@@ -8,7 +8,7 @@ public class CTJoinPacket() : StreamPacket(CTOffsets.CTJoinPacket)
 {
     public override void Read(PacketStream stream)
     {
-        var accountId = stream.ReadUInt32();
+        var accountId = stream.ReadUInt64();
         var cookie = stream.ReadUInt32();
 
         StreamManager.Instance.Login(Connection, accountId, cookie);
