+++ AAEmu.Game/Core/Packets/L2G/LGPlayerEnterPacket.cs
@@ -8,7 +8,7 @@ public class LGPlayerEnterPacket() : LoginPacket(LGOffsets.LGPlayerEnterPacket)
 {
     public override void Read(PacketStream stream)
     {
-        var accountId = stream.ReadUInt32();
+        var accountId = stream.ReadUInt64();
         var connectionId = stream.ReadUInt32();
         EnterWorldManager.Instance.AddAccount(accountId, connectionId);
     }
